Allow plists in XML format
Some older files seem to use that format.

I have a file from early 2020 in my iCloud which has an XML plist, so it
seems plausible apple used to upload them in that format?

I obviously don't remember how I actually added that file or when
exactly, but you know, it's there.

Fixes #1059